Update auto-complete.js
Hi, 

I hope you can review this pull request.

It was useful for us to specify a different width based on the form width and not on the input field. 
This seemed a nice way to accomplice this, since the width is calculated and available via CSS>

This way we could override it specifically when needed. 

Thank you for taking the time to review this.